Repository input type updated

+++ src/EduTrack.Data/IRepositories/IRepository.cs
@@ -10,9 +10,9 @@ namespace EduTrack.Data.IRepositories
 {
     public interface IRepository<TEntity> where TEntity : Auditable
     {
-        Task<bool> DeleteAsync(long id);
+        Task<bool> DeleteAsync(int id);
         IQueryable<TEntity> SelectAll();
-        Task<TEntity> SelectByIdAsync(long id);
+        Task<TEntity> SelectByIdAsync(int id);
         Task<TEntity> SelectAsync(Expression<Func<TEntity, bool>> predicate);
         Task<TEntity> InsertAsync(TEntity entity);
         Task<TEntity> UpdateAsync(TEntity entity);
